aursync: read updates via mapfile and lastpipe
Command substitution in an array is subject to word splitting, unlike
readarray (mapfile) which uses newline as a delimiter. As we wish to
preserve the exit status of aurcheck, use lastpipe instead of process
substitution (Bash 4.2). Shellcheck is not aware of this, so disable
SC2030/2031. An alternative would be wait "$!" (Bash 4.4), or writing
to a temporary file.