aurbuild: drop pool support
Support for a different pool location was introduced because of #129.
However, since then it was clear that the database files could simply
live in a cache directory, not requiring a separation. repose further
defaults to the value of --root if the --pool parameter is not given.

This allows to make repo-add optional with little code bloat, reading
the built packages into an array with null separation (the -d option to
mapfile requires bash 4.4; less elegant options are possible with
earlier bash versions)