rpc-deps, src-deps -> deps-rpc, deps-src
Scripts such as build-nspawn and fetch-snapshot are natural extensions
of build and fetch, respectively, that don't rely on the parent program.

rpc-deps on the other hand is mostly unrelated to rpc, apart from using
it to retrieve the necessary dependency information. As such, rename it
to deps-rpc.

Rename src-deps to deps-src as well, as the goal is identical to
deps-src but using a different method (local files instead of a remote
interface).