vercmp: add -u/upair
The goals were the following:

1) Do not rely on non-empty input (#273), neither for stdin nor the
   comparison (aur or file).
2) For -d to not make unnecessary version checks; this relies on a short
   join output. aurcheck (1.5.3) had 25k lines of (bash -x) output when
   checking the community repository against the AUR.
3) When using -c, if packages from stdin are not available in the repo,
   format "n/a -> version" with the cmp_equal_or_newer function.

To achieve this, it seems either multiple pipelines are required, or
different files to print unpairable lines from (join -a). "-o auto" was
avoided as format, since it fails with goal 1).

The idea is that -u sets "which file we are interested in"; when
comparing a repo against the AUR, the output from the AUR is what
matters. Thus, join -a 1 with a pipeline where this information appears
as the first file.

Conversely, when interested about the data from stdin (such as when
comparing versions with cmp_equal_or_newer) in a pipeline where this
information appears as the second file, use join -a 2.

The format (join -o) is fixed, but the file for unpaired lines can be
freely chosen from the command line with "vercmp -u".