Create update-devel
This example script is based on `aur-sync` changes proposed in #283. 

It is a simple reuse of `aurutils` components, similar to the examples in `aur(1)`. As such it is part of a new `examples` directory. This directory could include other scripts such as `aur-remove` (#305) and `aur-gc` (#317).

To use a script from the examples directory, it may be either copied to or symlinked in a directory in `PATH`. A symlink should be preferable unless local changes to the file are prefered.

 $ cp -s /usr/share/aurutils/examples/aur-update-devel /usr/local/bin
 $ aur update-devel