aur-build: account for makepkg --sign
If signing is enabled, re-use existing signatures from makepkg --sign
if this argument is specified on the command line (after --) or
enabled in BUILDENV (makepkg.conf). Note that if signing is disabled,
these signatures are ignored and only the package file is moved to the
repository root.

Some other miscellaneous changes.