aur-sync: introduce AUR_VIEWDIR
Patch files are only shown before AUR changes are merged. On
subsequent invocations of aur-sync, they are ignored, even if the user
did not approve the changes (for example, using :cq with vifm).

Using a dedicated directory (defined in the AUR_VIEWDIR environment
variable) allows to preserve the patches for later inspection.

See #379