add `--dump-options` to all commands using `parseopt.sh`
The main motivation behind this flag it to provide a list of options
that should be used in autocomplete scripts.

It should output to stdout one option per line including
the `-` or `--` prefix in the case of short and long opts respectively,
and the `:` suffix for options where an argument is required.

As regular users won't be using `--dump-options` directly, add a new
`opt_hidden` array were options that should not be listed by
`--dump-options` can be defined.

Here, not only "secret" options like `--dump-options` can added but also
the alternative forms of some options e.g. `--nover` vs `--no-ver`.
Effectively preventing redundant options from showing in completions.

The latter, however, is to be done in a future commit.