aur-repo: simplify pacman query, add -S
Because repository names cannot contain newlines, pacsift --null is not
required and the output can be parsed line-by-line. For a simple command
line, expac is used instead of pacman --print.

Add a sync "modifier" that applies to both package listing and package
upgrades. It is disabled by default; in particular, aur repo -u is now
equivalent to "aur repo -l | aur vercmp". If the repository path is
remote, hint to -S which enables it.

-Sl and -l have the same output: pkgname and pkgver separated by tabs.