aur-repo-filter: first steps into supporting versioned dependencies
As it turns out, pacsift --satisfies supports versioned strings
foo>=1.23, but the '>' should be properly quoted or it is taken as a
shell redirect. Due to odd issues with xargs/parallel, go back to
regular arguments (ARG_MAX is reached at 70k packages, which should be
more than sufficient).

TODO: Use versioned provides/replaces in expac and compare them.