aur: Don't add lib_dir to PATH if already there
The aur wrapper prepends $lib_dir to $PATH to ensure it is using its own
scripts. But due to the modular nature of aurutils the wrapper is called
multiple times throughout multiple pipelines causing multiple instances
of $lib_dir being added to $PATH.

Only prepend $lib_dir to $PATH if it isn't the first occurrence.