fetch: simplify "pull --rebase" fail check
`git pull --rebase` already exits with error when it fails to apply the
changes. This means that our current check is unreachable and
unnecessary.

`git` itself already prints advice on how to proceed, so let's simplify
the check and only print a small message to make a bit more clear which
package caused the failure.