fetch: change log format from git log to git diff
`aur-fetch` generates a "patch" based on the output of `git log`, giving
the user an overview of incoming changes.

But if the user doesn't update often enough, those patches would grow as
more commits are added, often just consisting of $pkgver bumps and
.SRCINFO updates, or showing intermediate changes that are no longer
relevant, making it harder to understand what ultimately changed.

To make review easier, show the incoming changes as single patch.