aur-fetch: add --rebase, --reset
This commit adds the ability to choose between git reset (default,
discard local commits) and git pull --rebase (attempt to reapply local
commits on top of incoming changes) to update cached repositories.

The behavior can either be set for all packages on the command-line with
--reset or --rebase, or per-repository with the aurutils.rebase
git-config(1) option.

The --force option and modified fetchspec was designed as a safer way to
update when we planned to make the --rebase behaviour default. We
reverted that decision, keeping --reset behaviour from 2.3.0, but
offering --rebase as an option to the user. Users of --rebase should be
familiar with git-rebase(1) as this option assumes manual intervention
may be required.