sync: do not save state information in vifminfo
To help with review we call vifm with some commands via -c.
Depending on what the user has vifminfo set to, those changes may be
made persistent between sessions.

Since some of the options are only enabled for the specific usecase of
reviewing pkgbuilds, and to avoid cloberring the userstate, unset
vifminfo to prevent these changes from being written to the vifminfo
file.