tests: verify parseopt loop consistency
Adding or renaming options is a two step process. The option needs to be
specified in the variables that are processed by parseopts and again in
the while true loop where we associate the options with their purpose.

A couple of regressions were caused by this, either directly because one
of the aforementioned steps was skipped or indirectly by incorrectly
solving conflicts on merge/rebase.

To detect mismatches, we parse the loop and compare the patterns
specified with the output of --dump-options (which reports the options
set in the variables).

To ensure we also cover opt_hidden mismatches we also teach dump-options
to return hidden options when AUR_DEBUG is set.

Allowing us to verify that all options in variables have an associated
case) in the loop and vice-versa.