actions: run shellcheck
If a step from a job fails the remaining steps are cancelled.
But we are interested in the results from shellcheck, even when the
option parsing consistency test fails.

For that, we include the condition: `if: success() || failure()` which
ensures the step will run even when the job status is already failure.