completion: fix broken completion for aur repo
When we attempt to generate a completion we check if the previous 'word'
matches 'word:' because options which require arguments are marked with
an ':' suffix in the array.

If we find a match, we complete with files by default,
otherwise we show all available options for the subcommand being used.

This worked until aur-repo learned '--repo:'.  Making aur repo <tab>
complete with files, as the subcommand itself matched one of its
options with required arguments.

To prevent that, always complete with options if we are on the second
word.