aur-build: use aur-repo semantics for repository selection
Instead of keeping a second repository selection logic in aur-build
(with slight differences to the one from aur-repo), simply forward the
-d and -r arguments to aur-repo, unless both are set. In particular,
aur-build can now be used with only the -r argument set.

This is achieved with the new aur-repo --status option (colon-delimited
output of repository name, path and root), which replaces aur repo --path.