aur-repo: explicitly declare dependencies as array
In case a repository package has no dependencies (i.e. the %DEPENDS%
pattern is not executed), the `dependencies` variable is not initialized
as an array.

The next package will then match %FILENAME%, and execute
`length(dependencies) > 0` which initializes `dependencies` as a scalar
(with value 0).

Avoid this by explicitly defining `dependencies` in BEGIN. Additionally,
avoid defining `dependencies` as a scalar in the first place by removing
`length(dependencies)`.

If --table is set, every package now is listed as depending on itself.
This matches the behavior of aur-graph and aur-depends. Strictly
speaking, this is only required for packages with no dependencies (so
they still appear when piping to tsort.)

Fixes #813