sync: filter repo contents seperately
When done in a pipe (and without set -o pipefail), the pipeline

  aur repo --status-file=db ... | complement_pattern ...

will use the exit status of complement_pattern. Specifically, if an
invalid repository is specified with -d, aur-sync will try to read the
`db` file even when it was not printed. This results in the error

  repo: invalid: repository path not found
  /usr/lib/aurutils/aur-sync: line 268: db: No such file or directory

Filter the repository contents in two steps instead.