query: reduce default of AUR_QUERY_RPC_SPLITNO_POST to 500
The aurweb configuration [1] specifies 5000 maximum rpc results, while
the empirical limit with aur-query is 730 packages. Whether this is a
curl (--parallel), aurweb or aur-query bug is yet to be determined.

Reduce the default number of AUR_QUERY_RPC_SPLITNO_POST to 500 as a
workaround. The case of querying more than 100 packages with aur-query
or aur-search should be rare, either way.

Closes #920

[1] https://gitlab.archlinux.org/archlinux/aurweb/-/blob/master/conf/config.defaults#L24