sync: build packages with ninja (EXPERIMENTAL)
Instead of keeping track of dependencies with a queue file:

depends1
depends2
depends3 ...

this is now done with a graph:

pkgbase1    depends1
pkgbase1    depends2
pkgbase1    depends3    ...

which is in turn fed to aur-sync--ninja. The generated build.ninja file
is used to resolve dependencies, and build packages by-branch (a failure
in a branch will not effect packages in other, independent branches).

Issue #908