ninja: take source directory as argument
The idea is to have the following directory layout

AURDEST
|- pkgbase1
|- pkgbase2
|- ...
|- sync-<identifier>
|-- build.ninja
|-- pkgbase1.stamp
|-- pkgbase2.stamp
|-- ...

This avoids aur-sync processes treading on eachother's toes, and *.stamp
files spreading out in $AURDEST. Specifically, with a unique "build"
directory per aur-sync invocation, `ninja -t clean` is not required to
build the full dependency tree.

The build directory is stored directly under AURDEST, such that it can
easily be recovered (for later calls of ninja outside of aur-sync)
or shared (e.g. in github repositories).

To make this work with general paths, aur-sync--ninja takes a directory
as its first argument, which is escaped according to ninja rules.