query: avoid hardcoding request types
When considering a "default" query type, we can make two observations:

1. search-type requests work both with HTTP POST and HTTP GET:

  $ curl -d 'v=5' -d type='search' -d by='name' -d 'arg=aurutils' https://aur.archlinux.org/rpc
  {"resultcount":2,"results": ...

2. aurweb has error-handling for invalid or missing type parameters:

  $ curl -d 'v=5' -d type='foobar' -d by='name' -d 'arg=aurutils' https://aur.archlinux.org/rpc
  {"error":"Incorrect request type specified.","resultcount":0,"results":[],"type":"error","version":5}

It thus makes sense to decouple GET and POST methods from the specific
type of request. This way, a wrong (or missing) type is directly
handled by aurweb. In addition, the default "by" search argument from
aurweb no longer needs to be hardcoded in aur-query.

The approach is not as general as it could be, because the POST and GET
functions still hardcode the existence of a "type" and "by" argument.
Instead, a key=value (similar to curl -d) approach could be taken.

Search requests now use HTTP POST to make the AUR_QUERY_RPC_POST
option more consistent, however this needs further testing.

Issue #938