aur-sync.1: use paclist instead of aur-vercmp-devel
In the Notes section, aur-vercmp-devel is used to demonstrate aur-sync
pipelines. Use the more general (and arguably more common) paclist as
example instead.

Instead of redirecting to a file, mention process substitution and command
substitution, both which can be done in a single command. Splitting on
white space is not a concern here, since pkgname cannot contain spaces.