chroot: exit if --create or --update was specified
Command-line arguments serve as makechrootpkg options with --build, and
as package targets with --create. --update does not use command-line
arguments, but occurs after --create, and before --build.

Exit when --create or --update was specified to avoid issues because of
this. A proper solution is to use --margs for makechrootpkg arguments in
--build, and keep positional arguments for package targets (Issue #807).

Issue #950