chroot: use positional arguments for packages
The `aur-chroot` command-line was changed so that `--create`, `--build` and
`--update` can be used in a single command. Command-line arguments are now
exclusively used for packages passed to `mkarchroot` and `arch-nspawn`.

To set `makechrootpkg` options, the new `--margs` and `--cargs` options are
available. As `aur-build --margs`, these take a comma-delimited string and
can be specified multiple times.