build: use makechrootpkg -c as default (--chroot)
For the longest time, aur-build would first update the container
template (<directory>/root), and then update the working copy again with
each build (<directory>/$USER) through `makechrootpkg -cu`.

If, in terms of linked dependencies, there is a practical difference
between -c and -cu is yet to be determined: in Arch, large rebuilds are
published to [staging] over a course of multiple days, instead of being
pushed to [testing]. In a single transaction, it is thus unlikely that
packages link against differently versioned dependencies. If only the
container template is updated at the start of the transaction, though,
and an update for a dependency was pushed somewhere in the middle, this
would be noticed with a 404 error.

ninja-based builds also use a single aur-build call per package,
resulting with -cu in two pacman upgrades per package. With -c, it is
one pacman upgrade. As such, one may argue the new behavior is
inconsistent: with makechrootpkg -c, regular builds have 1 pacman
upgrade per transaction, and ninja builds 1 build per package. This
could be remedied by making the first ninja target call `aur chroot
--update`, and have remaining `aur-build` command lines not call this
command. Per the above, there is unlikely to be a practical difference
either way (famous last words...)