build: allow secondary repositories for local repo sysupgrade
When a local repository package is upgraded, it may receive new
dependencies that are part of a different repository. For example, gmsh
added the community/cgns dependency with 4.9.5-2. If this package is not
installed, aur-build aborts with:

  warning: cannot resolve "cgns", a dependency of "gmsh"
  The following package cannot be upgraded due to unresolvable dependencies:
    gmsh

This happens because the pacman -Syu command in aur-build has a
configuration which only contains the local repository. This is done to
upgrade build dependencies, while avoiding an unattended -Syu of the
full system.

The solution is to use pacman 4.2's `Usage` parameters, with `Usage =
All` for the local repository (allowing a refresh) and `Usage = Install`
for all remaining repositories.

Issue #956