add examples directory
To make updating and copying of sample scripts more convenient, place
them in an `examples` directory. For now this contains the
`aur-sync-devel` script from aurvcs.7. The subdirectory contains a
`README.md` with a basic overview of the included scripts.

This is superficially similar to the previous `contrib` directory, but
emphasizes the use as examples, instead of "unsupported" third-party
scripts.