aur-sync: redirect aur-fetch output to stderr
In aur-sync, the --no-build option prints the paths of target packages.

With the recent addition of the `git log` check in aur-fetch, the output
of `aur sync --no-build` has now become polluted with the `aur fetch`
output.
That makes the output of `aur sync --no-build` less usable when
automating builds.

Fix that by redirecting the `aur-fetch` output (which is purely
diagnostic anyway) to standard error.