fetch: --existing: clone instead of skip AUR targets
The main goal of `--existing` is to disregard targets which are not in
AUR (i.e. not clone empty repositories). However, skipping clones
entirely is one step too far: dependencies may be added to AUR packages
over time, and we want to clone those if they are present in AUR.

The description in the man page is the same (and arguably the new
behavior matches it better).