fetch: check for revisions explicitly
Check if git-fetch wrote FETCH_HEAD with git rev-parse --verify -q.
This prints no error message, and exits 1 if FETCH_HEAD is not
available. When so, print a warning.

Remove the git-log command for existing packages as it is arguably only
useful when aur-fetch performs no destructive changes (git-rebase, git-reset).