lib: replace PACMAN_AUTH with AUR_PACMAN_AUTH
PACMAN_AUTH, in the context of makepkg, is a bash array set in
/etc/makepkg.conf. aur-build and aur-chroot assume it to be an
environment variable (split on whitespace) which cannot be an array. cf.

https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap08.html

Make this clear by defining the elevation command with AUR_PACMAN_AUTH.
When using makepkg --syncdeps (aur build --syncdeps) and host builds,
users should set both PACMAN_AUTH in makepkg.conf, and AUR_PACMAN_AUTH
in the aurutils environment.