pkglist: add --systime
Using date(1) directly to compare the packagelist timestamp depends on a
functioning/synchronized clock.

If the clock gets behind (e.g. in a "paused" virtual machine), and
date(1) is used to compare the time stamps, `sec_d - sec_l` can be
negative (or falsely below `ttl`) resulting in a delayed list update.

To address this, make the new behavior opt-in, i.e. when using
aur-pkglist for tab completions.