build: sign files in $var_tmp
When packages are built, `gpg` is run in `$var_tmp`. This was not the
case when packages are existing in the local repository, but their
signatures are not. Running `gpg` in `$var_tmp` also in this case
ensures:

- the build user (--user) has write access to the signature;
- `gpg` does not operate directly on the local repository;
- `gpg` does not try to overwrite signatures (in case of concurrent processes).

The multiple cases that need to be considered are formulated in the
small proof below. Note that a build user still needs read access to
packages in the local repository (implicit from `as_user repo-add`).

%<---------
let $PWD == $var_tmp, $db_root != $var_tmp,
    $p local repository path $db_root/<package>, $p_base = basename $p

assumption: --sign not specified to makepkg (otherwise, $var_tmp already contains a signature)

package existing, signature in local repo
we have:
  ! -f $p_base.sig    => ! -f $p_base.sig   [preserve signature]
    -f $p.sig
  ! -f $p_base
    -f $p

package existing, no signature in local repo
we have:
  ! -f $p_base.sig    =>   -f $p_base.sig   [write new signature]
  ! -f $p.sig
  ! -f $p_base
    -f $p

new package was built, basename identical to entry in local repo (--rebuild)
we have:
  ! -f $p_base.sig    =>   -f $p_base.sig
    -f $p.sig
    -f $p_base
    -f $p

new package was built, basename does not exist in local repo
we have:
  ! -f $p_base.sig    =>   -f $p_base.sig
  ! -f $p.sig
    -f $p_base
  ! -f $p