lib: add aur-repo-parse
repo-parse is a generalization of aur-repo --table, converting pacman
repositories into a JSON representation. `repo` options like `--attr`,
`--list` and `--list-attr` are also supported.

The current design stores the database contents into a single JSON array
to allow easy piping into aur-format (formatting strings). This is
unlike `repod`, which stores a single JSON array for each `pkgbase`.

Due to various problems with library `tar` implementations, the output
of `/bin/tar` is parsed in-order. It is assumed that `FILENAME` will
occur as the first field.

Performance note: Databases are not stored in memory, but processed
iteratively. The main consideration here are larger `.files` databases.