lib: add aur-format
This utility allows to custom-format searches, much like `cower` and
`package-query`. The heavy lifting is done in perl, and replaces use of
jq(1) in the code base.

* aur-format --table for tabulated output in aur-depends;
* set intersection and union for aur-search;
* aur-format --list for aur-vercmp.

The design is kept general and can both be used with RPCv5 output and
JSON arrays, such as the aurweb metadata dumps. Formatting keys are, to
the extent possible, modeled after package-query(1).

With the (optional) JSON::XS perl module, performance is good enough for
repeated processing of larger inputs. For example:

Benchmark 1: aur format packages-meta-ext-v1.json -f '%n\t%v'
  Time (mean ± σ):     907.9 ms ±   5.5 ms    [User: 808.8 ms, System: 98.5 ms]
  Range (min … max):   900.9 ms … 918.9 ms    10 runs

To perform set operations (intersection/union), all inputs are however
stored in memory.