fetch: introduce AUR_FETCH_USE_MIRROR
Remote clones are not migrated to local clones and vice-versa. This
process has some subtleties (such as checking if master@{u} is still
available) and is therefore left to the user.

To better reflect this, use an environment variable AUR_FETCH_USE_MIRROR
as a more permanent way to enable local clones than a command-line option.