repo-filter: store pacsift results in memory
Closing `pacsift` stdin when used in a pipe is not always reliable, with
one report mentioning it has no effect with QEMU+ssh. Either
`repo-filter` can revert to using `unbuffer`, or store the results in
memory. Since `unbuffer` or `expected` is not used in any other places,
go with the latter approach.