build--sync: escalate commands separately with sudo
Instead of running the whole script with sudo (either by invoking
`sudo aur build--sync ...` or running `exec sudo ...` inside the
script), run separate commands (pacsync, pacman) with sudo. This follows
the approach in aur-chroot.

This results in a potentially higher number of prompts when a low
timeout is configured in `sudoers(5)`, but leads to a more transparent
`sudoers(5)` configuration.