view-delta: revert to git-delta
The issues in 6ea812ef4013a3554e807a7e26bd1527465c1970 are addressed by
disabling side-by-side view and enabling --diff-so-fancy. In the master
branch of git-delta there also seem no more issues with exiting the
pager.

The main benefit of git-delta (compared to diff-so-fancy) is the ability
to enable line numbers.