build--pkglist: add --user
To avoid sourcing PKGBUILDs as root via aur-build--pkglist with a
privileged aur-build, one can:

1. define AUR_PKGLIST in aur-build and point it to a command that drops
   privileges, such as setpriv ... build--pkglist;
2. drop privileges to a specified user with exec setpriv ... directly
   inside build--pkglist.

The first option is similar to AUR_MAKEPKG, AUR_GPG and AUR_REPO_ADD.
All three are executables which are outside the control of aur-build.

aur-build--pkglist however can define its own behavior and drop
privileges automatically when a user is specified as argument (similar
to makechrootpkg -U). Due to the narrow scope of aur-build--pkglist this
approach seems preferable overall.