pkglist: revert to .gz archives
Using the git metadata archives is significantly slower than retrieving the
metadata archives, even with git-fetch --depth=1. It also uses an order of
magnitude more disk space, unless git-gc is run on a regular basis.

Those wanting to experiment with the git archives can either revert this commit
or resort to the pkglist-git branch.