lib: add aur-search--helper
Splitting out the aur-search awk code has the following advantages:

* `--format-stdin` can directly take JSON output with no loss, e.g.
  for `aur pkglist -i | aur search` pipelines;
* implementation details are centralized in one helper (locale, exit
  codes, escape sequences, language for text processing).

`aur-search` then remains for option processing and constructing a
pipeline from aur-format and aur-search--helper components.

One nuisance of keeping locale details as an `env` shebang in
`aur-search--helper` is that `file` no longer recognizes the source as
`text/x-awk` but as `text/plain`. `shellcheck` also no longer recognizes
the shebang, and so an exception was added to the Makefile.