aur-search: perl rewrite
This commit moves logic from `aur-format` to the higher-level
`aur-search`. This includes set intersection and union of results.
`aur-search` generally only deals with small inputs (AUR responses), so
loading all input into memory is not an issue here.

The scope of `aur-query` was not increased. This means `aur-search`
implements set intersections directly. The `--json` output now matches
`aur-pkglist`, as an array of hashes.

The `-r` short flag now points to `--reverse`, to match `aur depends -r`
and `pactree -r`. Previously it would point to `--raw` / `--json`.