view: remove diag_invalid
In the general case, aur-view only handles commits for a fixed AUR
repository. An invalid commit should therefore only happen when the
history was written by a TU. diag_invalid() on the other hand suggested
a problem with the "upstream URL".

These kind of ambiguities are better left in the man page, instead of a
fixed warning in aur-view.