build: revert e045ca35374095389288cc3a55657e7fbd4f1799
The main drawback is that a fixed pkgrel increase is assumed for every
PKGBUILD. Specifically, if the PKGBUILD for one package is restored, a
higher pkgrel increment is required so that `pacman -Syu` installs the
rebuilt package.

A script that wraps both aur-repo and aur-build could address this by
resetting pkgrel to the local repository version, or not touch it at all
if pkgver differs between local repo and PKGBUILD.

This may be implemented in the future as an example script
`sync-rebuild` which wraps `aur-sync --no-build`, or an `aur-rebuild`
called in conjunction with `aur-sync` or `aur sync--ninja`.