depends: add --assume-installed
RequiredBy gives for every node in the dependency graph all predecessor
node that depend on it. The `--assume-installed` flag removes all
specified targets from the graph, as well as from all `RequiredBy`
arrays. This is done by scanning the targets for every reverse dependency.

When afterwards a node has an empty `RequiredBy` array, it is considered
as no longer required and removed from the dependency graph. As a
result, dependencies for `--assume-installed` targets are also removed
from the graph. All this is done after `provides` are solved (`chain_mod`)
unless `--no-provides` is supplied.

A direct use for this is expanding `aur-sync --ignore` to also remove
dependencies of ignored targets. Later changes could use a similar
mechanism to remove transitive dependencies for `aur-repo-filter`
targets (issue #592).