build: do not use host for chroot configuration
In commit e670b2057cfb3cdde3dfa1451835f3938afcaa98, the host configuration was
used to set a fallback value for the chroot `pacman.conf`. This assumes that the
container has a local repository configured that is also configured on the host,
which is not necessarily true.

To still have a valid fallback as described in aur-build(1), only use
`/etc/aurutils/pacman-$db_name.conf` if `$db_name` is set explicitly through
`--database` or `AUR_REPO`.