depends: do not print undefined targets
When querying without `--show-all`, `aur-depends` assumes all targets
have the `PackageBase` field available. However, this only holds for
targets specified on the command-line when they are either AUR targets
themselves, or provided by a differently named AUR target.

When this does not hold (for example, the command-line is in the Arch
repositories), `aur-depends` would still try to print `PackageBase`,
resulting in empty fields for the `tsort` input.

To solve the problem at hand, make no special distinctions for
`--show-all` and check whether `PackageBase` is defined directly when
printing.

Closes #1062