Json.pm: return hash references
If a hash reference is encountered without `results` member, it can
either be the output from `aur-repo-parse --jsonl` or the git metadata
archives.

Since latter were removed and the former is useful to reduce memory
consumption when processing large databases with `aur-format`, make
`parse_json_aur()` return the hash reference directly in this case.