repo: do not require a configured file:// repository
The configured repository can be remote (http or https). In this case,
the `pacman-conf` loop will set $db_root to `Server` (if `--root` is
not specified) or to the value of `--root` otherwise. Because of the
latter, there is no hard requirement on `Server` being `file://`.

The difference with <=17.3 is when the repository is not configured,
with a valid `--root` given:

  $ aur repo -d ewfwefew --root /home/custompkgs  # 17.3
  repo: /home/custompkgs/ewfwefew.db: not a file

  $ aur repo -d ewfwefew --root /home/custompkgs  # this commit
  repo: ewfwefew: repository not configured

When the repository was created, but not configured:

  $ aur repo -d custom --root /home/custompkgs    # 17.3
  repo:custom
  root:/home/custompkgs
  path:/home/custompkgs/custom.db.tar.gz

  $ aur repo -d custom --root /home/custompkgs    # this commit
  repo: custom: repository not configured

Issue #1113