sync: include every target with `select_pkgbase()`
Commit ba7a9601aff2f3077d331f4069a02c2538ba1d22 removes `Self` lines for
dependencies in `aur-depends` which are not specified as targets to its
command-line. This way transitive dependencies for `--assume-installed`
targets are removed from the output.

If an `aur-sync` target is already in the local repository, but recieves
additional dependencies (e.g. from building with --no-check to building
regularly), this change implies the latter dependencies would not be
included. (Issue #1136)

Solve it this time around by modifying the `select_pkgbase()` function
to print all targets.