repo: remove explicit file/directory checks
Consider a script `aur-repo-init`, which aims to create a local
repository by the file paths defined in pacman.conf. With a
file/directory check in place, `aur-repo` would fail to print the
necessary paths, even though the repository is not directly accessed.

Remove the checks to address this use case. Necessary file checks are
still done by `aur-repo-parse`, when actually reading from the local
repository. The same holds for `aur-build` for write access.

Reported-by: Ivan Shapovalov <intelfx@intelfx.name>